+++ db/migrate/20210208200031_rename_esr_participant_nr_to_esr_account_in_organisation.rb
@@ -1,0 +1,6 @@
+class RenameESRParticipantNrToESRAccountInOrganisation < ActiveRecord::Migration[6.1]
+  def change
+    rename_column :organisations, :esr_participant_nr, :esr_beneficiary_account
+    add_column :organisations, :esr_ref_prefix, :string, null: true
+  end
+end