+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+- Add prefix for ESR reference numbers
+- Rename ESR beneficiary account
+
 ## [21.2.1] - 2021-02-07
 
 - Improve checkbox actions to be more direct